+++ pkg/common/common.go
@@ -1,0 +1,9 @@
+package common
+
+import "github.com/giantswarm/logging-operator/pkg/key"
+
+func IsLoggingEnabled(labels map[string]string) bool {
+	value, ok := labels[key.LoggingLabel]
+
+	return ok && value == "true"
+}