+++ pkg/key/key.go
@@ -1,0 +1,6 @@
+package key
+
+const (
+	Finalizer    = "logging-operator.finalizers.giantswarm.io"
+	LoggingLabel = "giantswarm.io/logging"
+)